+++ milkman/src/main/java/milkman/ui/main/keys/KeyResolverPlugin.java
@@ -2,10 +2,14 @@ package milkman.ui.main.keys;
 
 import lombok.Setter;
 import milkman.domain.KeySet;
+import milkman.domain.KeySet.KeyEntry;
 import milkman.ui.main.Toaster;
 import milkman.ui.plugin.ActiveKeySetAware;
 import milkman.ui.plugin.TemplateParameterResolverPlugin;
 import milkman.ui.plugin.ToasterAware;
+
+import java.util.List;
+import java.util.stream.Collectors;
 
 public class KeyResolverPlugin implements TemplateParameterResolverPlugin, ActiveKeySetAware, ToasterAware {
 
@@ -30,4 +34,10 @@ public class KeyResolverPlugin implements TemplateParameterResolverPlugin, Activ
         return "key";
     }
 
+    @Override
+    public List<String> getAllEntries() {
+        return activeKeySet.getEntries().stream()
+                .map(KeyEntry::getName)
+                .collect(Collectors.toList());
+    }
 }