+++ milkman/src/main/java/milkman/ui/plugin/TemplateParameterResolverPlugin.java
@@ -1,5 +1,7 @@
 package milkman.ui.plugin;
 
+import java.util.List;
+
 /**
 * a templater that modifies input-strings in an
 * implementation-specific way, e.g. replaces environment-varables.
@@ -9,12 +11,16 @@ public interface TemplateParameterResolverPlugin {
     /**
      * returns a string with all replaced tags
      */
-    public String lookupValue(String input);
+    String lookupValue(String input);
 
     /**
      * the prefix under which the templater is registered
      * @return
      */
-    public String getPrefix();
+    String getPrefix();
+
+    default List<String> getAllEntries(){
+        return List.of();
+    }
 
 }