+++ milkman/src/main/java/milkman/ctrl/ExecutionListenerManager.java
@@ -2,7 +2,7 @@ package milkman.ctrl;
 
 import milkman.domain.RequestContainer;
 import milkman.domain.ResponseContainer;
-import milkman.ui.plugin.ExecutionListenerAware;
+import milkman.ui.plugin.ExecutionListenerAware.ExecutionListener;
 
 import javax.inject.Singleton;
 import java.util.HashMap;
@@ -11,9 +11,9 @@ import java.util.Optional;
 
 @Singleton
 public class ExecutionListenerManager {
-	private final Map<String, Map<String, ExecutionListenerAware.ExecutionListener>> listeners = new HashMap<>();
+	private final Map<String, Map<String, ExecutionListener>> listeners = new HashMap<>();
 
-	public void listenOnExecution(RequestContainer request, String listenerId, ExecutionListenerAware.ExecutionListener listener) {
+	public void listenOnExecution(RequestContainer request, String listenerId, ExecutionListener listener) {
 		var requestListeners = listeners.computeIfAbsent(request.getId(), id -> new HashMap<>());
 		requestListeners.put(listenerId, listener);
 	}
@@ -23,6 +23,13 @@ public class ExecutionListenerManager {
 				.map(Map::values)
 				.ifPresent(listeners ->
 						listeners.forEach(l -> l.onRequestStarted(request, response)));
+	}
+
+	/* package */ void onRequestReady(RequestContainer request, ResponseContainer response) {
+		Optional.ofNullable(listeners.get(request.getId()))
+				.map(Map::values)
+				.ifPresent(listeners ->
+						listeners.forEach(l -> l.onRequestReady(request, response)));
 	}
 
 	/* package */ void onRequestFinished(RequestContainer request, ResponseContainer response) {