+++ changelog.md
@@ -11,6 +11,7 @@ version 5.3.0 -- upcoming
   * mask plain-keys on overview
   * timeouts for oauth
   * auto-refreh for oauth only on usage
+  * auto-completion now includes prefixed resolvers (such as `key:`)
 
 version 5.2.0
   * added key-store that is not synced