+++ changelog.md
@@ -12,6 +12,7 @@ version 5.3.0 -- upcoming
   * timeouts for oauth
   * auto-refreh for oauth only on usage
   * auto-completion now includes prefixed resolvers (such as `key:`)
+  * new [WebSocket plugin](/milkman-ws)
 
 version 5.2.0
   * added key-store that is not synced